ebl  Tweak code to compile win64 version with mingw
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@8584 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/win32/filed/vss.cpp
@@ -215,7 +215,7 @@ const char* VSSClient::GetWriterInfo(int nIndex)
 const int VSSClient::GetWriterState(int nIndex)
 {
    alist* pV = (alist*)m_pAlistWriterState;   
-   return (int)pV->get(nIndex);
+   return (intptr_t)pV->get(nIndex);
 }
 
 void VSSClient::AppendWriterInfo(int nState, const char* pszInfo)
